+++ empire/server/modules/powershell/management/spawn.py
@@ -22,24 +22,35 @@ class Module(object):
         proxy = params["Proxy"]
         proxy_creds = params["ProxyCreds"]
         sys_wow64 = params["SysWow64"]
+        language = params["Language"]
+
         if (params["Obfuscate"]).lower() == "true":
             launcher_obfuscate = True
         else:
             launcher_obfuscate = False
         launcher_obfuscate_command = params["ObfuscateCommand"]
 
-        # generate the launcher script
-        launcher = main_menu.stagers.generate_launcher(
-            listenerName=listener_name,
-            language="powershell",
-            encode=True,
-            obfuscate=launcher_obfuscate,
-            obfuscation_command=launcher_obfuscate_command,
-            userAgent=user_agent,
-            proxy=proxy,
-            proxyCreds=proxy_creds,
-            bypasses=params["Bypasses"],
-        )
+        if language == "powershell":
+            # generate the launcher script
+            launcher = main_menu.stagers.generate_launcher(
+                listenerName=listener_name,
+                language=language,
+                encode=True,
+                obfuscate=launcher_obfuscate,
+                obfuscation_command=launcher_obfuscate_command,
+                userAgent=user_agent,
+                proxy=proxy,
+                proxyCreds=proxy_creds,
+                bypasses=params["Bypasses"],
+            )
+        elif language in ["csharp", "ironpython"]:
+            launcher = main_menu.stagers.generate_exe_oneliner(
+                language=language,
+                obfuscate=obfuscate,
+                obfuscation_command=launcher_obfuscate,
+                encode=True,
+                listener_name=listener_name,
+            )
 
         if launcher == "":
             return handle_error_message("[!] Error in launcher command generation.")